benchmark/pyproject.toml
[tool.poetry]
name = "cdle"
version = "0.1.0"
description = ""
authors = ["Robert Gleison"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
joblib = "^1.4.2"
modin = "^0.32.0"
pyspark = "^3.5.5"
dask = {extras = ["distributed"], version = "^2025.4.1"}
pyarrow = "^20.0.0"
numpy = "^1.23"
gcsfs = "^2024.5.0"  
dask-kubernetes = "^2025.4.3"
snakeviz = "^2.2.2"
python-dotenv = "^1.1.0"
scikit-learn = "^1.6.1"
matplotlib = "^3.10.3"
xgboost = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
